+++ src/index.ts
@@ -1,12 +1,16 @@
 export { ThemeProvider, ThemeConsumer } from "./context";
-export { default as useTheme } from "./useTheme";
-export { default as withStyles } from "./withStyles";
-export { default as makeUseStyles, MakeUseStylesFn } from "./makeUseStyles";
-export { default as useColorTransition } from "./useColorTransition";
+export { useTheme } from "./useTheme";
+export { withStyles } from "./withStyles";
+export { makeUseStyles, MakeUseStylesFn } from "./makeUseStyles";
+export { useColorTransition } from "./useColorTransition";
+export { orientation } from "./orientation";
 export {
   maxWidth,
   maxHeight,
   minHeight,
   minWidth,
+  aspectRatio,
+  maxAspectRatio,
+  minAspectRatio,
   createDimensionQueryHelper
 } from "./mediaQueryHelpers";