+++ src/makeUseStyles.d.ts
@@ -3,20 +3,16 @@ import { StyleSheet } from "react-native";
 export type MakeUseStylesFn<Theme extends {}> = <
   T extends StyleSheet.NamedStyles<T> | StyleSheet.NamedStyles<any>
 >(
-  styles:
+  getStyles:
     | ((theme: Theme) => T | StyleSheet.NamedStyles<T>)
-    | T
-    | StyleSheet.NamedStyles<T>
+    | (() => T | StyleSheet.NamedStyles<T>)
 ) => () => T;
 
-declare function makeUseStyles<
+export function makeUseStyles<
   Theme extends {},
   T extends StyleSheet.NamedStyles<T> | StyleSheet.NamedStyles<any>
 >(
-  styles:
+    getStyles:
     | ((theme: Theme) => T | StyleSheet.NamedStyles<T>)
-    | T
-    | StyleSheet.NamedStyles<T>
+    | (() => T | StyleSheet.NamedStyles<T>)
 ): () => T;
-
-export default makeUseStyles;