+++ src/makeUseStyles.js
@@ -1,33 +1,21 @@
 import { useMemo } from "react";
 import { StyleSheet } from "react-native";
 
-import resolveMediaQueries from "./resolveMediaQueries";
 import useDimensions from "./useDimensions";
-import useTheme from "./useTheme";
+import { useTheme } from "./useTheme";
 
-const makeUseStyles = styles => {
+export const makeUseStyles = getStyles => {
+  const hasThemeDependency = getStyles.length === 0;
   const useStyles = () => {
-    const isDynamic = typeof styles === "function";
     const theme = useTheme();
-    const { window: windowDimensions } = useDimensions();
 
-    return useMemo(
-      () =>
-        StyleSheet.create(
-          resolveMediaQueries(
-            isDynamic ? styles(theme) : styles,
-            windowDimensions
-          )
-        ),
-      [
-        // do not use theme as dependency when passed styled are not function
-        isDynamic ? theme : null, 
-        windowDimensions
-      ]
-    );
+    return useMemo(() => StyleSheet.create(getStyles(theme)), [
+      // `windowDimensions` need to media query helpers (orientation, maxHeight. ...) update styles
+      useDimensions().window,
+      // Don't use `theme` as dependency when passed styled are not function
+      hasThemeDependency ? theme : null
+    ]);
   };
 
   return useStyles;
 };
-
-export default makeUseStyles;