+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 2.0.0
+
+Reduce bundle size
+Deprecate first argument as object `makeUseStyles({ styles: {color: 'red'}})` now only functions allow `makeUseStyles (() => (({ styles: {color : 'red'}}))`  
+Added new media query helpers:
+
+- `aspectRatio(ratio: number,{/* styles */})`
+- `maxAspectRatio(ratio: number,{/* styles */})`
+- `minAspectRatio(ratio: number{/* styles */})`
+- `orientation({ portrait: {},landscape: {} })`
+
 # 1.1.3
 
 - Improve recreating styles